Relocate verifier excerpt cluster to verifier_orchestration
Phase 2 batch 4 for Issue #682. Moves the verifier file-excerpt
cluster (~196 LOC) out of turn.rs into verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  Pure helpers:
    - verifier_diagnostic_file_excerpts (33 LOC)
    - verifier_repair_context_line_for_path (11 LOC)
    - safe_verifier_diagnostic_file_excerpt (34 LOC, includes
      canonical-path symlink containment + mask_secrets pipeline +
      control-char neutralize for Issue #638 Task 1.8 + Codex CB-002)
    - verifier_file_excerpt_for_line (28 LOC)
    - build_verifier_excerpt (20 LOC)
    - head_tail_excerpt (13 LOC)

  Struct + impl:
    - VerifierExcerptWindow + 6 method impl (51 LOC), now
      pub(super) with its `::new` constructor also pub(super); the
      remaining impl-block helpers stay private to the impl block
      (only used by sibling free fns above).

  Consts:
    - VERIFIER_DIAGNOSTIC_MAX_FILE_EXCERPTS (cap = 6)
    - VERIFIER_DIAGNOSTIC_MAX_FILE_EXCERPT_BYTES (cap = 1400)

All items promoted to pub(super). The struct + 2 consts had no
external references outside the moved cluster, so they moved
cleanly.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::progress_text::truncate
    - super::repair_framework_findings::VerifierDiagnosticFileExcerpt
    - super::repair_job::mask_secrets_headers_and_neutralize
    - super::verifier_repair_targeting::verifier_diagnostic_path_input_is_safe
    - std::collections::HashSet

  * turn.rs:
    - Extended use super::verifier_orchestration::{...} with the
      3 production-callable relocated fns
      (verifier_diagnostic_file_excerpts /
      verifier_file_excerpt_for_line /
      verifier_repair_context_line_for_path).
    - Removed VerifierDiagnosticFileExcerpt from the production
      use super::repair_framework_findings import; re-added under
      #[cfg(test)] so the test-only callers (~7 sites that build
      VerifierDiagnosticFileExcerpt literals) still resolve.
    - Removed verifier_diagnostic_path_input_is_safe from the
      verifier_repair_targeting use (only the moved fn references
      it).
    - Cleaned an orphaned #[derive(Debug, Clone, Copy)] attribute
      and 4 trailing blank lines that were left behind when the
      VerifierExcerptWindow struct was extracted.

Metrics: turn.rs 34,561 -> 34,353 (-208); verifier_orchestration.rs
484 -> 697 (+213). Cumulative Issue #682 progress:
turn.rs 34,958 -> 34,353 (-605, ~21% of Phase 2 target). Acceptance
gap: 34,353 -> 32,000 = ~2,353 LOC remaining.
